requirements.txt
# Dependencias principales
python-binance>=1.0.16
pandas>=1.3.0
numpy>=1.20.0
python-dotenv>=0.19.0

# Visualización (opcional)
matplotlib>=3.4.0
seaborn>=0.11.0

# Análisis de datos
ta>=0.7.0  # Biblioteca de indicadores técnicos

# Utilidades
requests>=2.25.0
websocket-client>=1.2.0

# Desarrollo y testing
pytest>=6.2.0
pytest-cov>=2.12.0 